+++ doc/release_notes.md
@@ -17,6 +17,8 @@ commands which show an explanation about the command.
 - Expressions with strings in PKSMacroC such as "x"+1 are automatically casted to string. One does not need to write any more "x"+(string)1.
 - PKSMacroC supports now `++` and `--` operators and the new datatype `float`.
 - One can now debug print the low level instructions of PKSMacroC macros (from the macro dialog).
+- /* */ comments in macros are now supported and automatically attached to a macro as a comment (e.g. used as help in code completion), 
+  if placed in front of a macro definition.
 
 ### Changed
 - all files previously stored in the old Windows INI-file format (`pksedit.his`, `pksedit.ini`) were replaced by JSON formatted files: